Overhaul our multilingual support.
You can now extract text for just one language, or every language in the file.

This also fixes a number of problems we have with nested language and non-language templates; which was the primary motivating factor for this refactor. Being able to pull the whole language set is just a nice bonus.