Finish implementing data augmentation for label fields in CLIP.
This requires some extra finangling, because it turns out I *never actually tested* the prior changes I did to enable wikitext parsing. This is due to the caching in Datasets.

Datasets chokes on "ragged" datasets, so I have to load the whole set and then insert emtpy strings for values that don't exist. No, we can't use defaultdict, datasets bypasses that.

At some point we need to move to SQL, because the flat filesystem data model is killing us.