Allow scraping category data as well as file articles.
The changes in this PR includes fixes for a number of strange corner-cases that MediaWiki induces: namely that a category can be referenced without actually having a corresponding page, resulting in MediaWiki generating ad-hoc IDs just to say "this page doesn't exist". Thus we hack in some extra hacks into the merge dict code that powers querying.

Originally this was just going to grab all parent categories of each scraped file, but that would result in the application scraping tens of thousands of categories for a handful of images.